Migrate builders golden check metrics into golden library (#8191)
### Ticket
Closes [#8189](https://github.com/tenstorrent/tt-mlir/issues/8189) 

### Problem description
The functions used to calculate golden checks are located in the builder
module instead of golden. Having them in golden allows for shared use
between chisel and builder minimizing code duplication and maintaining
consistency

### What's changed
Migrated `mask_torch_inf_nan` and `get_atol_rtol_pcc` to new file
`tools/golden/metrics.py`

### Checklist
- [ ] New/Existing tests provide coverage for changes

---------
