+++ tools/golden/__init__.py
@@ -2,10 +2,14 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 from .mapping import *
+from .metrics import *
 
 __all__ = [
     "GoldenMapTensor",
     "unpack_mlir_attr",
     "get_golden_function",
     "GOLDEN_MAPPINGS",
+    "get_pcc",
+    "get_atol_rtol",
+    "get_atol_rtol_pcc",
 ]
